requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
sqlparse==0.5.3
